+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
 #### Added
 - nothing yet
 
@@ -11,6 +12,46 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 #### Changed
 - nothing yet
+
+## [3.5.3]
+#### Fixed
+- Fixed an [issue](https://github.com/Iterable/react-native-sdk/issues/547) where the SDK would crash if the `IterableInAppMessage` object was null when consuming an in-app message.
+
+#### Changed
+- SDK now provides more insight into JWT refresh failures, so you can take appropriate action in your application code. With this update, when a JWT refresh fails (for any  of various reasons), the SDK calls `onAuthFailure` on the `IterableAuthHandler` instance you can provide to the SDK at initialization.
+    For more information, see the [documentation](https://support.iterable.com/hc/articles/360035019712) for Iterable's Android SDK.
+- The  IterableAuthHandler's `onTokenRegistrationFailed(Throwable object)` has been replaced with `onAuthFailure(Throwable object)` which will be called when JWT token refresh fails.
+
+## [3.5.2]
+#### Added
+ - Introducing a new method `setAuthRetryPolicy` in `IterableConfig` to set the retry policy for JWT token refresh.
+ - You can now use method - `pauseAuthRetries(boolean pauseRetry)` to pause or resume the JWT token refresh retries.
+ - `trackPurchase` method now accepts `attributionInfo` parameter to track purchase with attribution information.
+
+#### Fixed
+- Embedded manager listeners are active only if feature is enabled
+
+## [3.5.1]
+
+#### Fixed
+- Embedded manager syncs messages as soon as user logs in
+- Fixed a bug to prevent app crashes due to malformed authTokens
+- Fixes a bug where push notifications opens would lead to app crashes
+- Fixes a bug where crashes would occur due to SDK not being initialized
+
+#### Changed
+- Updates Glide library in embedded message OOTB views to version [4.16.0](https://github.com/bumptech/glide/releases/tag/v4.16.0)
+
+## [3.5.0]
+#### Added
+- introduces support for embedded messaging: an eligibility–based, personalized messages sent from Iterable to your mobile and web apps, which can display them inline, using native interface components.
+- To display embedded messages, you can use customizable, out-of-the-box components provided by the SDK (cards, notifications, banners), or you can build fully custom components of your own design.
+- To learn more, read [Embedded Messages with Iterable's iOS SDK](https://support.iterable.com/hc/articles/23061840746900).
+
+#### Changed
+- `IterableConfig` is updated with an `enableEmbeddedMessaging` flag that needs to be set to true to allow use of embedded messaging functionality
+
+## [3.5.0-beta1](https://github.com/Iterable/iterable-android-sdk/releases/tag/3.5.0-beta1)
 
 ## [3.4.17](https://github.com/Iterable/iterable-android-sdk/releases/tag/3.4.17)
 #### Added